add a Sys module and start to move some system constants out of base (also closes #3331)
