+++ base/exports.jl
@@ -13,6 +13,7 @@ export
     GMP,
     QuadGK,
     Sort,
+    Sys,
     Test,
     Pkg,
     Operators,
