+++ base/util.jl
@@ -227,20 +227,20 @@ function versioninfo(io::IO, verbose::Bool)
     println(io,             "Julia $version_string")
     println(io,             commit_string)
     println(io,             "Platform Info:")
-    println(io,             "  OS_NAME: ", OS_NAME)
-    println(io,             "  WORD_SIZE: ", WORD_SIZE)
+    println(io,             "  System: ", Sys.OS_NAME, " (", Sys.MACHINE, ")")
+    println(io,             "  WORD_SIZE: ", Sys.WORD_SIZE)
     if verbose
         lsb = readchomp(ignorestatus(`lsb_release -ds`) .> SpawnNullStream())
         if lsb != ""
             println(io,     "           ", lsb)
         end
         println(io,         "  uname: ",readchomp(`uname -mprsv`))
-        println(io,         "Memory: $(total_memory()/2^30) GB ($(free_memory()/2^20) MB free)")
-        try println(io,     "Uptime: $(uptime()) sec") catch end
+        println(io,         "Memory: $(Sys.total_memory()/2^30) GB ($(Sys.free_memory()/2^20) MB free)")
+        try println(io,     "Uptime: $(Sys.uptime()) sec") catch end
         print(io,           "Load Avg: ")
-        print_matrix(io,    Base.loadavg()')
+        print_matrix(io,    Sys.loadavg()')
         println(io          )
-        println(io,         cpu_info())
+        println(io,         Sys.cpu_info())
     end
     if Base.libblas_name == "libopenblas"
         openblas_config = openblas_get_config()
@@ -258,117 +258,11 @@ function versioninfo(io::IO, verbose::Bool)
             end
         end
         println(io          )
+        println(io,         "Package Directory: ", Pkg.dir())
         println(io,         "Packages Installed:")
         Pkg.status(io       )
     end
 end
-
-type UV_cpu_info_t
-    model::Ptr{Uint8}
-    speed::Int32
-    cpu_times!user::Uint64
-    cpu_times!nice::Uint64
-    cpu_times!sys::Uint64
-    cpu_times!idle::Uint64
-    cpu_times!irq::Uint64
-end
-type CPUinfo
-    model::ASCIIString
-    speed::Int32
-    cpu_times!user::Uint64
-    cpu_times!nice::Uint64
-    cpu_times!sys::Uint64
-    cpu_times!idle::Uint64
-    cpu_times!irq::Uint64
-    SC_CLK_TCK::Int
-    CPUinfo(model,speed,u,n,s,id,ir,ticks)=new(model,speed,u,n,s,id,ir,ticks)
-end
-CPUinfo(info::UV_cpu_info_t, ticks) = CPUinfo(bytestring(info.model), info.speed,
-    info.cpu_times!user, info.cpu_times!nice, info.cpu_times!sys,
-    info.cpu_times!idle, info.cpu_times!irq, ticks)
-
-show(io::IO, cpu::CPUinfo) = show(io, cpu, true, "    ")
-function show(io::IO, info::CPUinfo, header::Bool, prefix::String)
-    tck = info.SC_CLK_TCK 
-    if header
-        println(io, info.model, ": ")
-        print(" "^length(prefix))
-        if tck > 0
-            @printf io "    %5s    %9s    %9s    %9s    %9s    %9s\n" "speed" "user" "nice" "sys" "idle" "irq"
-        else
-            @printf io "    %5s    %9s  %9s  %9s  %9s  %9s ticks\n" "speed" "user" "nice" "sys" "idle" "irq"
-        end
-    end
-    print(prefix)
-    if tck > 0
-        @printf io "%5d MHz  %9d s  %9d s  %9d s  %9d s  %9d s" info.speed info.cpu_times!user/tck info.cpu_times!nice/tck info.cpu_times!sys/tck info.cpu_times!idle/tck info.cpu_times!irq/tck
-    else
-        @printf io "%5d MHz  %9d  %9d  %9d  %9d  %9d ticks" info.speed info.cpu_times!user info.cpu_times!nice info.cpu_times!sys info.cpu_times!idle info.cpu_times!irq
-    end
-end
-function cpu_summary(io::IO, cpu::Array{CPUinfo}, i, j)
-    if j-i < 9
-        header = true
-        for x = i:j
-            if header == false println() end
-            show(io,cpu[x],header,"#$(x-i+1) ")
-            header = false
-        end
-    else
-        summary = CPUinfo(cpu[i].model,0,0,0,0,0,0,cpu[i].SC_CLK_TCK)
-        count = j-i+1
-        for x = i:j
-            summary.speed += cpu[i].speed
-            summary.cpu_times!user += cpu[x].cpu_times!user
-            summary.cpu_times!nice += cpu[x].cpu_times!nice
-            summary.cpu_times!sys += cpu[x].cpu_times!sys
-            summary.cpu_times!idle += cpu[x].cpu_times!idle
-            summary.cpu_times!irq += cpu[x].cpu_times!irq
-        end
-        summary.speed = div(summary.speed,count)
-        show(io,summary,true,"#1-$(count) ") 
-    end
-end
-function show(io::IO, cpu::Array{CPUinfo})
-    model = cpu[1].model
-    first = 1
-    for i = 2:length(cpu)
-        if model != cpu[i].model
-            cpu_summary(io,cpu,first,i-1)
-            first = i
-        end
-    end
-    cpu_summary(io,cpu,first,length(cpu))
-end
-repl_show(io::IO, cpu::Array{CPUinfo}) = show(io, cpu)
-function cpu_info()
-    SC_CLK_TCK = ccall(:jl_SC_CLK_TCK, Clong, ())
-    UVcpus = Array(Ptr{UV_cpu_info_t},1)
-    count = Array(Int32,1)
-    uv_error("uv_cpu_info",ccall(:uv_cpu_info, UV_error_t, (Ptr{Ptr{UV_cpu_info_t}}, Ptr{Int32}), UVcpus, count))
-    cpus = Array(CPUinfo,count[1])
-    for i = 1:length(cpus)
-        cpus[i] = CPUinfo(unsafe_load(UVcpus[1],i),SC_CLK_TCK)
-    end
-    ccall(:uv_free_cpu_info, Void, (Ptr{UV_cpu_info_t}, Int32), UVcpus[1], count[1])
-    cpus
-end
-
-function uptime()
-    uptime_ = Array(Float64,1)
-    uv_error("uv_uptime",ccall(:uv_uptime, UV_error_t, (Ptr{Float64},), uptime_))
-    return uptime_[1]
-end
-
-function loadavg()
-    loadavg_ = Array(Float64,3)
-    ccall(:uv_loadavg, Void, (Ptr{Float64},), loadavg_)
-    return loadavg_
-end
-
-free_memory() = ccall(:uv_get_free_memory, Uint64, ())
-total_memory() = ccall(:uv_get_total_memory, Uint64, ())
-
 
 # `methodswith` -- shows a list of methods using the type given
